test: Simplify TagFilterTest, move detail cases to TagFilterMatcher
Start to simplify the TagFilterTest by moving many of the edge cases
and repeated detail cases to the TagFilterMatcherTest. We can start to
boil TagFilterTest down to cases that specifically show how the filter
applies across a Feature / Scenario / Example hierarchy, rather than
the detail of the filter string syntax.
